+++ app/models/invoice/e_invoice_generator.rb
@@ -41,22 +41,16 @@ class Invoice
 
       e_invoice_invoice_items = []
       invoice.each do |invoice_item|
-        e_invoice_invoice_item = EInvoice::InvoiceItem.new.tap do |i|
-          i.description = invoice_item.description
-          i.price = invoice_item.price
-          i.quantity = invoice_item.quantity
-          i.unit = invoice_item.unit
-          i.subtotal = invoice_item.subtotal
-          i.vat_rate = invoice_item.vat_rate
-          i.vat_amount = invoice_item.vat_amount
-          i.total = invoice_item.total
-        end
+        e_invoice_invoice_item = generate_invoice_item(invoice, invoice_item)
         e_invoice_invoice_items << e_invoice_invoice_item
       end
+
+      e_invoice_name_item = e_invoice_invoice_items.shift if invoice.monthly_invoice
 
       e_invoice_invoice = EInvoice::Invoice.new.tap do |i|
         i.seller = seller
         i.buyer = buyer
+        i.name = e_invoice_name_item&.description
         i.items = e_invoice_invoice_items
         i.number = invoice.number
         i.date = invoice.issue_date
@@ -72,9 +66,33 @@ class Invoice
         i.currency = invoice.currency
         i.delivery_channel = %i[internet_bank portal]
         i.payable = payable
+        i.monthly_invoice = invoice.monthly_invoice
       end
 
       EInvoice::EInvoice.new(date: Time.zone.today, invoice: e_invoice_invoice)
+    end
+
+    private
+
+    def generate_invoice_item(invoice, item)
+      EInvoice::InvoiceItem.new.tap do |i|
+        i.description = item.description
+        i.unit = item.unit
+        i.price = item.price
+        i.quantity = item.quantity
+        if invoice.monthly_invoice && item.price && item.quantity
+          i.product_id = item.product_id
+          i.vat_rate = invoice.vat_rate
+          i.subtotal = (item.price * item.quantity).round(3)
+          i.vat_amount = i.subtotal * (i.vat_rate / 100)
+          i.total = i.subtotal + i.vat_amount
+        else
+          i.subtotal = item.subtotal
+          i.vat_rate = item.vat_rate
+          i.vat_amount = item.vat_amount
+          i.total = item.total
+        end
+      end
     end
   end
 end