+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+02.09.2022
+* Update invoice status on payment order payments [#2427](https://github.com/internetee/registry/pull/2427)
+
+01.09.2022
+* Monthly invoice payment status fix [#2428](https://github.com/internetee/registry/issues/2428)
+
+31.08.2022
+* new fully automated process for registrar monthly invoices [#2424](https://github.com/internetee/registry/pull/2424)
+
 25.08.2022
 * Contact creation fix to not require postal addresses in Registrar portal [#2421](https://github.com/internetee/registry/pull/2421)